feat(redesign): Removing buildSrc directory
