+++ src/app/build.gradle.kts
@@ -1,6 +1,13 @@
 @file:Suppress("UnstableApiUsage")
 @Suppress("DSL_SCOPE_VIOLATION")
 
+fun Project.debugAPIAuth() = findProperty("MOVIE_DB_API_TOKEN_DEBUG")
+    ?: System.getenv("MOVIE_DB_API_TOKEN_DEBUG")
+
+fun Project.releaseAPIAuth() = findProperty("MOVIE_DB_API_TOKEN_RELEASE")
+    ?: System.getenv("MOVIE_DB_API_TOKEN_RELEASE")
+
+
 plugins {
     id("com.android.application")
     kotlin("android")
