--- src/buildSrc/build.gradle.kts
@@ -1,10 +1,0 @@
-import org.gradle.kotlin.dsl.`kotlin-dsl`
-
-plugins {
-    `kotlin-dsl`
-    `kotlin-dsl-precompiled-script-plugins`
-}
-
-repositories {
-    mavenCentral()
-}
