--- src/buildSrc/src/main/java/ProjectExt.kt
@@ -1,8 +1,0 @@
-import org.gradle.api.Project
-
-
-fun Project.debugAPIAuth() = findProperty("MOVIE_DB_API_TOKEN_DEBUG")
-    ?: System.getenv("MOVIE_DB_API_TOKEN_DEBUG")
-
-fun Project.releaseAPIAuth() = findProperty("MOVIE_DB_API_TOKEN_RELEASE")
-    ?: System.getenv("MOVIE_DB_API_TOKEN_RELEASE")
